requirements.txt
absl-py==2.4.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.61.1
gensim==4.3.3
h5py==3.15.1
hmmlearn==0.3.3
ImageIO==2.37.2
joblib==1.5.3
keras==3.9.0
kiwisolver==1.4.9
lazy_loader==0.4
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
networkx==3.4.2
nlkt==4.0.1
nltk==3.9.2
numpy==1.26.4
opencv-python==4.13.0.92
optree==0.18.0
packaging==26.0
pandas==2.2.3
pillow==12.1.1
Pygments==2.19.2
pyparsing==3.3.2
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2026.1.15
rich==14.3.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.13.1
six==1.17.0
sklearn-crfsuite==0.5.0
smart_open==7.5.0
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
wrapt==2.1.1